# public/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "version": "1.0.0",
        "description": "Documentation of API endpoints for the application"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Serveur Local"
        }
    ],"components": {
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        
        }
    },
    "security": [
        {
            "BearerAuth": []
    
            }
        ],
    "paths": {
...
        "/api/v1/referentiels": {
            "get": {
                "summary": "Get all active referentiels",
                "description": "This endpoint is used to retrieve a list of active referentiels.",
                "operationId": "getActiveReferentiels",
                "tags": [
                    "Referentiels"
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object",
                                        "properties": {
                                            "id": {
                                                "type": "string"
                                            },
                                            "libelle": {
                                                "type": "string"
                                            },
                                            "etat": {
                                                "type": "string"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/v1/referentiel": {
            "post": {
                "summary": "Create a new referentiel",
                "description": "This endpoint is used to create a new referentiel.",
                "operationId": "createReferentiel",
                "tags": [
                    "Referentiels"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "libelle": {
                                        "type": "string"
                                    },
                                    "etat": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "id": {
                                            "type": "string"
                                        },
                                        "libelle": {
                                            "type": "string"
                                        },
                                        "etat": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/v1/referentiels/etat": {
            "get": {
                "summary": "Get referentiels by state",
                "description": "This endpoint is used to retrieve referentiels by their state.",
                "operationId": "getReferentielsByEtat",
                "tags": [
                    "Referentiels"
                ],
                "parameters": [
                    {
                        "name": "etat",
                        "in": "query",
                        "description": "Filter referentiels by state",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object",
                                        "properties": {
                                            "id": {
                                                "type": "string"
                                            },
                                            "libelle": {
                                                "type": "string"
                                            },
                                            "etat": {
                                                "type": "string"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/v1/referentiels/{id}/competences": {
            "get": {
                "summary": "Get referentiel competences",
                "description": "This endpoint is used to retrieve competences of a referentiel.",
                "operationId": "getReferentielCompetences",
                "tags": [
                    "Referentiels"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Referentiel ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object",
                                        "properties": {
                                            "id": {
                                                "type": "string"
                                            },
                                            "libelle": {
                                                "type": "string"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/v1/referentiels/{id}/modules": {
            "get": {
                "summary": "Get referentiel modules",
                "description": "This endpoint is used to retrieve modules of a referentiel.",
                "operationId": "getReferentielModules",
                "tags": [
                    "Referentiels"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Referentiel ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object",
                                        "properties": {
                                            "id": {
                                                "type": "string"
                                            },
                                            "libelle": {
                                                "type": "string"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/v1/referentiels/{id}": {
            "patch": {
                "summary": "Update a referentiel",
                "description": "This endpoint is used to update a referentiel's information.",
                "operationId": "updateReferentiel",
                "tags": [
                    "Referentiels"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Referentiel ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "libelle": {
                                        "type": "string"
                                    },
                                    "etat": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "id": {
                                            "type": "string"
                                        },
                                        "libelle": {
                                            "type": "string"
                                        },
                                        "etat": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/v1/referentiel/{id}": {
            "delete": {
                "summary": "Soft delete a referentiel",
                "description": "This endpoint is used to soft delete a referentiel.",
                "operationId": "softDeleteReferentiel",
                "tags": [
                    "Referentiels"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Referentiel ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "id": {
                                            "type": "string"
                                        },
                                        "libelle": {
                                            "type": "string"
                                        },
                                        "etat": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/v1/referentiels/deleted": {
            "get": {
                "summary": "Get all deleted referentiels",
                "description": "This endpoint is used to retrieve a list of deleted referentiels.",
                "operationId": "getDeletedReferentiels",
                "tags": [
                    "Referentiels"
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object",
                                        "properties": {
                                            "id": {
                                                "type": "string"
                                            },
                                            "libelle": {
                                                "type": "string"
                                            },
                                            "etat": {
                                                "type": "string"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
...
  }
}